html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-color-01.html

<!DOCTYPE html>
<title>document: fg/bg/link/vlink/alink-color</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-fgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-body-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="iframe"></iframe>
<script>
test(function() {
  var doc = document.getElementById("iframe").contentDocument;
  var body = doc.documentElement.removeChild(doc.body);
  assert_equals(doc.fgColor, "");
  assert_equals(doc.bgColor, "");
  assert_equals(doc.linkColor, "");
  assert_equals(doc.vlinkColor, "");
  assert_equals(doc.alinkColor, "");
  doc.body = body;
})
</script>
